feat: separation of logic and move style to styles.tsx
